core/requirements.txt
python-telegram-bot
vk
requests
pony
flask